backend/requirements.txt
flask
faiss-cpu
requests
PyPDF2